feat: refactor ProgressEvent type definition for improved readability
